+++ com.unity.shadergraph/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - When you rapidly perform an undo or redo, Shader Graph windows no longer break.
 - Sub Graphs that contain references to non-existing Sub Graphs no longer break the Sub Graph Importer.
 - You can now reference sub-assets such as Textures.
-- SubGraph Output and CustomFunctionNode now validate slot names and display appropriate error badge.
+- Sub Graph Output and Custom Function Node now validate slot names, and display appropriate error badge when needed.
 
 ## [6.6.0] - 2019-04-01
 ### Added